# api/api-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IDClawserver API",
    "version": "0.1.0",
    "description": "API for AI agent identities backed by RODiT tokens on NEAR, including facial token_id encoding.",
    "x-policies": {
      "termsOfService": "/.well-known/terms-of-service",
      "privacyPolicy": "/.well-known/privacy-policy",
      "dataRetention": "/.well-known/data-retention"
    },
    "contact": {
      "name": "IDENTYCLAW Support",
      "email": "support@identityclaw.com",
      "url": "https://identityclaw.com/contact"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://identityclaw.com/license"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "RODiT-JWT",
        "description": "RODiT-issued JWT access token validated by @rodit/rodit-auth-be (RoditClient.authenticate), obtained via the RODiT login flow (e.g. /api/login). This is not an API key."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "nullable": true
              }
            },
            "required": ["code", "message"]
          },
          "requestId": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["error", "requestId", "timestamp"]
      }
    }
  },
  "paths": {
...
    "/api/agent/auth-params": {
      "get": {
        "summary": "Get timestamp and nonce for AI agent login",
        "description": "Public endpoint for AI agents to obtain timestamps and random nonces needed to construct /api/login requests. AI agents often struggle with real-time data and random number generation; this endpoint provides pre-generated values.",
        "tags": ["Authentication"],
        "responses": {
          "200": {
            "description": "Authentication parameters for constructing login requests.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "integer",
                      "description": "Unix timestamp in seconds (as expected by /api/login)."
                    },
                    "timestamp_iso": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO 8601 timestamp for reference."
                    },
                    "nonce": {
                      "type": "string",
                      "description": "Random 32-byte nonce encoded as base64url."
                    },
                    "nonce_hex": {
                      "type": "string",
                      "description": "Same nonce encoded as hexadecimal."
                    },
                    "nonce_length": {
                      "type": "integer",
                      "description": "Length of the nonce in bytes (32)."
                    },
                    "algorithm": {
                      "type": "string",
                      "description": "Algorithm used to generate the nonce."
                    },
                    "purpose": {
                      "type": "string",
                      "description": "Guidance on how to use these values."
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request correlation identifier."
                    }
                  },
                  "required": ["timestamp", "nonce", "nonce_length", "requestId"]
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - rate limit exceeded (100 requests per minute).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "rateLimit": {
                    "summary": "Rate limit exceeded",
                    "value": {
                      "error": {
                        "code": "RATE_LIMIT_EXCEEDED",
                        "message": "Exceeded anonymous auth-params rate limit"
                      },
                      "requestId": "01HX9WZ7F7B5DDXPKRZC4J8M0X",
                      "timestamp": "2026-04-15T08:20:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to generate authentication parameters.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "generationError": {
                    "summary": "Entropy generator failure",
                    "value": {
                      "error": {
                        "code": "AGENT_AUTH_PARAMS_FAILED",
                        "message": "Random generator unavailable"
                      },
                      "requestId": "01HX9WZ8TF5X1Q4P3Q2D0A9RZT",
                      "timestamp": "2026-04-15T08:20:02.000Z"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Get timestamp and nonce for AI agent login (POST alternative)",
        "description": "Alternative POST endpoint for AI agents that prefer POST requests. Returns the same data as the GET endpoint.",
        "tags": ["Authentication"],
        "responses": {
          "200": {
            "description": "Authentication parameters for constructing login requests.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "integer",
                      "description": "Unix timestamp in seconds (as expected by /api/login)."
                    },
                    "timestamp_iso": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO 8601 timestamp for reference."
                    },
                    "nonce": {
                      "type": "string",
                      "description": "Random 32-byte nonce encoded as base64url."
                    },
                    "nonce_hex": {
                      "type": "string",
                      "description": "Same nonce encoded as hexadecimal."
                    },
                    "nonce_length": {
                      "type": "integer",
                      "description": "Length of the nonce in bytes (32)."
                    },
                    "algorithm": {
                      "type": "string",
                      "description": "Algorithm used to generate the nonce."
                    },
                    "purpose": {
                      "type": "string",
                      "description": "Guidance on how to use these values."
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request correlation identifier."
                    }
                  },
                  "required": ["timestamp", "nonce", "nonce_length", "requestId"]
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - rate limit exceeded (100 requests per minute).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "rateLimit": {
                    "summary": "Rate limit exceeded",
                    "value": {
                      "error": {
                        "code": "RATE_LIMIT_EXCEEDED",
                        "message": "Exceeded anonymous auth-params rate limit"
                      },
                      "requestId": "01HX9WZ7F7B5DDXPKRZC4J8M0X",
                      "timestamp": "2026-04-15T08:20:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to generate authentication parameters.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "generationError": {
                    "summary": "Entropy generator failure",
                    "value": {
                      "error": {
                        "code": "AGENT_AUTH_PARAMS_FAILED",
                        "message": "Random generator unavailable"
                      },
                      "requestId": "01HX9WZ8TF5X1Q4P3Q2D0A9RZT",
                      "timestamp": "2026-04-15T08:20:02.000Z"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/agents": {
      "get": {
        "summary": "List RODiT token holders with facial descriptions",
        "description": "Returns a paginated list of all RODiT token holders with their facial descriptions decoded from token IDs. This is a public endpoint for agent discovery. For detailed identity info including DN and full metadata, use the protected /api/identity/token/{tokenId}/full endpoint.",
        "tags": ["Agent"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of agents to return (default: 20, max: 100)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor for fetching the next page",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of agents with facial descriptions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tokenId": {
                            "type": "string",
                            "description": "The 12-character facial token ID"
                          },
                          "face": {
                            "type": "object",
                            "nullable": true,
                            "description": "Facial description decoded from tokenId",
                            "properties": {
                              "checksumValid": {
                                "type": "boolean",
                                "description": "Whether the token ID checksum is valid"
                              },
                              "categories": {
                                "type": "object",
                                "description": "Facial feature categories with their values"
                              }
                            }
                          }
                        }
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true,
                      "description": "Cursor for fetching the next page, null if no more results"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request correlation identifier"
                    }
                  },
                  "required": ["agents", "requestId"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "500": {
            "description": "Failed to retrieve agents"
          }
        }
      }
    },
...
  }
}